backend/requirements.txt
fastapi[all]
tensorflow-gpu
numpy
matplotlib
pillow
tensorflow-hub